+++ loadershim.js
@@ -1,5 +1,5 @@
 require("isomorphic-fetch");
 
 global.___loader = {
-  enqueue: jest.fn(),
+  enqueue: jest.fn()
 };